+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     this in mind if you override them in your app.
 
 ### Grid
+*   The `del` action has been renamed to `delete`.
+
 *   Columns configuration (by using the `Grid.column` DSL method, or the `columns` configuration option) no longer has effect on corresponding form fields. Use the new `attribute` DSL method or the `attribute_overrides` config option (see above) to set the common configuration option for both column and form field.
 
 *   Virtual columns declared with the `column` DSL method are no longer automatically appended to the end of the column